requirements.txt
Django==1.9.1
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
PyYAML==3.11
Twisted==15.5.0
anyjson==0.3.3
appier==1.2.3
argh==0.26.1
argparse==1.2.1
cffi==1.5.0
characteristic==14.3.0
cryptography==1.2.1
django-cors-headers==1.1.0
djangorestframework==3.3.2
djangorestframework-gis==0.10.1
enum34==1.1.2
-e git+https://github.com/hendrix/hendrix.git@ece7d9b7d7fcc76d826d310eee043fad7360f5ee#egg=hendrix-master
idna==2.0
ipaddress==1.0.16
oauthlib==1.0.3
pathtools==0.1.2
psycopg2==2.6.1
pyOpenSSL==0.15.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pychalk==0.0.5
pycparser==2.14
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2015.7
requests==2.9.1
requests-oauth==0.4.1
requests-oauthlib==0.6.0
service-identity==14.0.0
simplejson==3.8.1
six==1.10.0
tweetstream==1.1.1
twitter-api==0.1.5
txsockjs==1.2.2
watchdog==0.8.3
wsgiref==0.1.2
zope.interface==4.1.3
